requirements.txt
streamlit>=1.30.0

langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-chroma>=0.1.0

chromadb>=0.4.22

sentence-transformers>=2.2.2

python-dotenv>=1.0.0

tavily-python>=0.3.3

pypdf>=3.17.0
python-docx>=1.1.0
